requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
imutils==0.5.4
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
opencv-contrib-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
seaborn==0.13.2
six==1.16.0
sympy==1.13.1
timm==1.0.11
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
